started handling unsafe code from the original
there are a few cute template tricks in the original code that quite
frankly feel pretty shitty, so I've decided to tackle those with more
explicit operands to the functions/macros that implement the
opcodes. This commit contains an obvious example of one of
these. Ideally the macro expansion/optimization in Rust should still
take care of these nicely.