fixed a logical-!-on-integer port TODO
I believe this to be correct now. Basically the original code takes a
value, applies a bitwise & to it with another value, then a logical
!. In C++ the int -> bool conversion is very simple; 0 is false,
anything else is true. The conversion back is also standard-conformant;
false goes to 0 and true to 1. As far as I know, Rust doesn't have
implicit int/bool conversions in either direction (which is a Good Thing
imo), so this has to be converted to something that works without the
logical ! operator. So basically the code is confusing but after some
thought I think it's correct. We'll have to see when I finish the port :)