lots more filling out of the run function
I realized I may have made another porting mistake. Some of the ops in
the c++ code use reference parameters, which I assumed might have been
some sort of speed optimization or something. I realize now, tho, that
this may not be the case; I'll need to go through the new ops and
replicate this, as it can definitely change behavior. For instance, if a
register is passed to one of these functions by reference, the
register's value might actually change before it's referenced in the new
function, and in that case we expect to use the updated value, not
necessarily the value when the function was called.

Ugh. This is very hairy code. I don't think I'll be baking any
bsnes/higan code anymore :P