Added APU function to clear the echo buffer
Many SPC files have garbage in the echo buffer on startup, so now
consumers of this library have the option to clear the echo buffer by
calling a function after setting the APU state from an SPC file. While
this isn't technically correct, many other SPC players also do
this (including blargg's, which is widely considered to be the most
accurate APU emulator around), so I think we're OK to do it too.

Fixes #5.