Fixed memory mapping bug
This was a bug that masked off certain memory addresses that many songs
would use to store data. Trivial fix, but very difficult to find :)

Fixes #31, which I consider to be the last known major bug in terms of
emulation (wooo!!)

Side note: the kirby intro track STILL crashes in this port when the
song finishes, but I found out it's because a WAI/SLEEP instruction is
executed at the end of the song (which is actually strange, as this
would lock up the entire audio unit on a real system.. perhaps it has
some kind of hard reset feature I don't know about). Anyways, I would
hardly consider that a bug at this point :)